+++ packages/pluggableWidgets/datagrid-web/src/utils/useShowPagination.ts
@@ -1,20 +1,28 @@
-import { useMemo } from "react";
+import { useMemo, useEffect } from "react";
 import { PaginationEnum, ShowPagingButtonsEnum } from "../../typings/DatagridProps";
 
 interface ShowPaginationProps {
     pagination: PaginationEnum;
-    showPagingButtons: ShowPagingButtonsEnum;
+    showPagingButtonsOrRows: ShowPagingButtonsEnum | boolean;
     totalCount?: number;
     limit: number;
+    requestTotalCount: (needTotalCount: boolean) => void;
 }
 
 export const useShowPagination = (props: ShowPaginationProps): boolean => {
-    const { pagination, showPagingButtons, totalCount, limit } = props;
+    const { pagination, showPagingButtonsOrRows, totalCount, limit, requestTotalCount } = props;
+
+    useEffect(() => {
+        if (pagination !== "buttons" && showPagingButtonsOrRows) {
+            requestTotalCount(true);
+        }
+    }, [pagination]);
+
     return useMemo(() => {
         return (
-            pagination === "buttons" &&
-            (showPagingButtons === "always" ||
-                (showPagingButtons === "auto" && (totalCount ? totalCount > limit : false)))
+            showPagingButtonsOrRows === true ||
+            showPagingButtonsOrRows === "always" ||
+            (showPagingButtonsOrRows === "auto" && (totalCount ? totalCount > limit : false))
         );
-    }, [pagination, showPagingButtons, totalCount, limit]);
+    }, [pagination, showPagingButtonsOrRows, totalCount, limit]);
 };