+++ packages/pluggableWidgets/datagrid-web/CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 
 -   We introduced a setting to exclude filters from being stored in the Personalization configuration.
+-   We have introduced show number of rows for virtual scrolling and load more pagination mode.
 
 ## [2.28.2] - 2024-12-12